+++ packages/shared/src/i18n/components/LocaleSuggest.astro
@@ -38,13 +38,13 @@ const currentLocale = Astro.currentLocale as Lang;
 >
   <div class="flex items-center gap-3">
     <div class="shrink-0">
-      <Icon name="tabler:language" class="size-5 text-emerald-400 dark:text-[#34d399]" />
+      <Icon name="tabler:language" class="size-5 text-emerald-700 dark:text-[#34d399]" />
     </div>
 
     <div class="grow">
       <p class="text-sm text-gray-800 dark:text-[#e2e8f0] font-medium">
         <span class="font-normal" id="js-suggestionText"></span>
-        <a id="js-link" href="#" class="text-emerald-600 dark:text-[#34d399] hover:text-emerald-700 dark:hover:text-emerald-300 underline-offset-2 hover:underline font-semibold transition-colors duration-200" aria-label="Cambiar idioma">
+        <a id="js-link" href="#" class="text-emerald-700 dark:text-[#34d399] hover:text-emerald-800 dark:hover:text-emerald-300 underline-offset-2 hover:underline font-semibold transition-colors duration-200" aria-label="Cambiar idioma">
           <span id="js-linkText"></span>
         </a>
       </p>
